I have optimized the performance by 10 times based on your new PR.Under the current test case, the performance of conv has improved from 0.62 GB/s to 6.47 GB/s. The performance of scan has increased from 0.30 GB/s to 2.86 GB/s
- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [ ] Medium
  - [ ] High